Update Fedor and New Game.cpp

+++ 1100/Fedor and New Game.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-	int n,m,k;
+	ll n,m,k;
 	cin>>n>>m>>k;
 	vector<int>a(10005);
 	for(int i=0;i<=m;i++)
